2007-02-22  Esteban Sanchez  <estebans@artica.es>
        * pandoradb.sql: Changed a 'utimestamp' field from mediumint to
        integer.


git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@382 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/pandoradb.sql
@@ -11,7 +11,7 @@ CREATE TABLE `tagent_access` (
   `id_ac` bigint(20) unsigned NOT NULL auto_increment,
   `id_agent` int(8) unsigned NOT NULL default '0',
   `timestamp` datetime NOT NULL default '0000-00-00 00:00:00',
-  `utimestamp` mediumint(12) unsigned NOT NULL default '0',
+  `utimestamp` integer unsigned NOT NULL default '0',
   PRIMARY KEY  (`id_ac`),
   KEY `agent_index` (`id_agent`)
 ) TYPE=InnoDB;
